git-svn-id: http://estimo.googlecode.com/svn/trunk@19 c8bd1b88-1bcb-11de-b69a-45dc1c5e4070

+++ x264/encoder/cas.c
@@ -135,7 +135,7 @@ cas_new(const uint32_t *prog_mem, const uint16_t *point_mem, int feu, int foptl,
     cas->last_qcycles = 0;
 
     cas->cycles = cas->fcycles = cas->qcycles = cas->could_save_cycles = 0;
-    cas->energy_pJ = 0.0;
+    cas->energy_pJ = (int64_t)0;
 
     return cas;
 }
